Reorganize Docker and conda
The docker file was not really efficient to iterate quickly.
Changing the order of things helps.

Also fixed the way we create the conda environment. Switched to a
python 2.7 environment as this is required for the verkiezingen use
case.

Finally, we need to activate the environment and start the app in
the same bash command.